fix wrong file reference
